iot_simulator/requirements.txt
mysql-connector-python==9.1.0
numpy==2.2.1
scipy==1.14.1
